chapter10/worker_service/requirements.txt
celery==3.1.23
kafka==1.3.5
requests==2.9.1
pillow==4.3.0
grpc
grpcio
google==1.9.3
tensorflow==1.3.0
tensorflow-serving-api==1.3.0
futures